% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muanova.h.R
\name{muANOVA}
\alias{muANOVA}
\title{Factorial ANOVA}
\usage{
muANOVA(
  data,
  dep,
  lev_fac_a = 3,
  lev_fac_b = 2,
  lev_fac_c = 2,
  type_fac_a = "b",
  type_fac_b = "b",
  type_fac_c = "b",
  n = 20,
  alpha = 0.05,
  stdev = 1,
  withcorr = 0,
  num_facs = "one"
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{lev_fac_a}{.}

\item{lev_fac_b}{.}

\item{lev_fac_c}{.}

\item{type_fac_a}{.}

\item{type_fac_b}{.}

\item{type_fac_c}{.}

\item{n}{.}

\item{alpha}{.}

\item{stdev}{.}

\item{withcorr}{.}

\item{num_facs}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$designtab} \tab \tab \tab \tab \tab a table \cr
  \code{results$main} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$designtab$asDF}

\code{as.data.frame(results$designtab)}
}
\description{
Factorial ANOVA
}
