% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestones.h.R
\name{ttestOneS}
\alias{ttestOneS}
\title{One Sample T-Test}
\usage{
ttestOneS(
  calc = "n",
  es = 0.5,
  power = 0.9,
  n = 20,
  alt = "two.sided",
  alpha = 0.05,
  powerContour = TRUE,
  powerDist = FALSE,
  powerCurveES = TRUE,
  powerCurveN = FALSE,
  text = TRUE
)
}
\arguments{
\item{calc}{.}

\item{es}{.}

\item{power}{.}

\item{n}{.}

\item{alt}{.}

\item{alpha}{.}

\item{powerContour}{.}

\item{powerDist}{.}

\item{powerCurveES}{.}

\item{powerCurveN}{.}

\item{text}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
One Sample T-Test
}
