% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestps.h.R
\name{ttestPS}
\alias{ttestPS}
\title{Paired Samples T-Test}
\usage{
ttestPS(
  calc = "n",
  es = 0.5,
  power = 0.9,
  n = 20,
  alt = "two.sided",
  alpha = 0.05,
  powerContour = TRUE,
  powerDist = FALSE,
  powerCurveES = TRUE,
  powerCurveN = FALSE,
  text = TRUE
)
}
\arguments{
\item{calc}{.}

\item{es}{.}

\item{power}{.}

\item{n}{.}

\item{alt}{.}

\item{alpha}{.}

\item{powerContour}{.}

\item{powerDist}{.}

\item{powerCurveES}{.}

\item{powerCurveN}{.}

\item{text}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$tabText} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerEStab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerContour} \tab \tab \tab \tab \tab an image \cr
  \code{results$contourText} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerCurveES} \tab \tab \tab \tab \tab an image \cr
  \code{results$curveESText} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerCurveN} \tab \tab \tab \tab \tab an image \cr
  \code{results$curveNText} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerDist} \tab \tab \tab \tab \tab an image \cr
  \code{results$distText} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Paired Samples T-Test
}
